Add subsystem for parsing / filtering tags defined from a single source
Allow the file name to be passed as an option, defaulting to a json defined at the project root.

The formatting of the JSON would be a key (tag_targets_mappings) that contains a dict where the keys are tag names, and the values are lists containing the target addresses to apply to.