Wire target tag definitions into the legacy parser
Worth noting - at this point, we're unable to distinguish much for targets at this point beyond the 'name' attribute, which may not be sufficient to uniquely match to the targets to apply tags to. Looking through GlobsHandlingTargetAdaptor and TargetAdaptor and was unable to locate more beyond that would help uniquely identify the appropriate target - something for further discussion. Additionally, we'll need to plug this into the V2 engine as well.